requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
geopy~=2.4.1
pillow~=11.0.0
pip~=24.3.1
attrs~=24.2.0
wheel~=0.37.1
PyYAML~=6.0.2
geographiclib~=2.0
asgiref~=3.8.1
sqlparse~=0.5.1
psycopg2~=2.9.10
referencing~=0.35.1
jsonschema~=4.23.0
setuptools~=60.2.0
uritemplate~=4.1.1
tzdata~=2024.2
inflection~=0.5.1
phonenumbers~=8.13.49
PyJWT~=2.9.0